Fix getWallet depends on address type / Start describe address domain (#3284 #3464)
