+++ src/common/domain/address.js
@@ -1,0 +1,16 @@
+export const AddressType = {
+  Internal: 'Internal',
+  Metamask: 'Metamask',
+  Custom: 'Custom',
+}
+
+export const AddressFormat = {
+  'Full': 'Full',
+  'Short': 'Short', // 0x12...cdef
+  'Auto': 'Auto', // todo
+}
+
+export const AddressRole = {
+  Send: 'Send',
+  Receive: 'Receive',
+}
