+++ src/front/shared/redux/actions/core.js
@@ -8,6 +8,7 @@ import Pair from 'pages/Exchange/Orders/Pair'
 import config from 'helpers/externalConfig'
 
 import metamask from 'helpers/metamask'
+import { AddressType } from 'domain/address'
 
 
 const debug = (...args) => console.log(...args)
@@ -343,7 +344,7 @@ const markCoinAsVisible = (coin, doBackup) => {
 
 const getWallet = (findCondition) => {
   // all wallets from all origins, including metamask
-  const { currency, address } = findCondition
+  const { currency, address, addressType } = findCondition
 
   const founded = getWallets().filter((wallet) => {
     const conditionOk = (currency && wallet.currency.toLowerCase() === currency.toLowerCase())
@@ -352,9 +353,17 @@ const getWallet = (findCondition) => {
       if (wallet.address.toLowerCase() === address.toLowerCase()) {
         return conditionOk
       }
-    } else {
-      return conditionOk
     }
+
+    if (addressType) {
+      if (addressType === AddressType.Internal && !wallet.isMetamask
+        || addressType === AddressType.Metamask && wallet.isMetamask
+      ) {
+        return conditionOk
+      }
+    }
+
+    return conditionOk
   })
 
   return (founded.length) ? founded[0] : false
